Use INTEGER for the score, not SMALLINT

+++ src/ingest_into_postgres_asyncpg.py
@@ -23,7 +23,7 @@ async def create_tables(conn):
             created_utc  TIMESTAMP WITH TIME ZONE,
             title        TEXT,
             retrieved_at TIMESTAMP WITH TIME ZONE,
-            score        SMALLINT,
+            score        INTEGER,
             permalink    TEXT,
             LOCKED       BOOLEAN,
             selftext     TEXT,
@@ -37,7 +37,7 @@ async def create_tables(conn):
             created_utc  TIMESTAMP WITH TIME ZONE,
             parent_id    TEXT,
             permalink    TEXT,
-            score        SMALLINT,
+            score        INTEGER,
             retrieved_at TIMESTAMP WITH TIME ZONE
         );
      """
