+++ src/ingest_into_postgres_psycopg3.py
@@ -21,7 +21,7 @@ def create_tables(conn):
             created_utc  TIMESTAMP WITH TIME ZONE,
             title        TEXT,
             retrieved_at TIMESTAMP WITH TIME ZONE,
-            score        SMALLINT,
+            score        INTEGER,
             permalink    TEXT,
             LOCKED       BOOLEAN,
             selftext     TEXT,
@@ -35,7 +35,7 @@ def create_tables(conn):
             created_utc  TIMESTAMP WITH TIME ZONE,
             parent_id    TEXT,
             permalink    TEXT,
-            score        SMALLINT,
+            score        INTEGER,
             retrieved_at TIMESTAMP WITH TIME ZONE
         );
      """
